Fixed some issues with the delete function

+++ src/components/Edits/AmiiboEditDelete.js
@@ -6,44 +6,57 @@ import "./AmiiboEditDelete.css";
 class EditDelete extends Component {
   constructor(props) {
     super(props);
-    this.state = { searchArray: [], editID: "", isLoading: false };
+    this.state = {
+      searchArray: [],
+      editID: "",
+      isLoading: false,
+      deleted: false
+    };
     this.data = {};
   }
 
   delete = event => {
-    event.preventDefault();
-    this.setState({ isLoading: true });
-    fetch(this.props.baseURL + "id/" + this.state.editID, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" }
-    })
-      .then(res => res.json())
-      .then(res => {
-        console.log("Got it!");
-        this.setState({
-          searchArray: res,
-          isLoading: false
-        });
-        // Sends res back to state in AmiiboSearch
-        this.props.setSearchArray(res);
+    if (!this.state.deleted && this.state.editID) {
+      event.preventDefault();
+      this.setState({ isLoading: true });
+      fetch(this.props.apiURL + "id/" + this.state.editID, {
+        method: "DELETE",
+        headers: { "Content-Type": "application/json" }
       })
-      .catch(err => {
-        console.log("We've got a problem, sir.", err);
-      });
+        .then(res => res.json())
+        .then(res => {
+          console.log("Got it!");
+          this.setState({
+            searchArray: res,
+            isLoading: false,
+            deleted: true
+          });
+          // Sends res back to state in AmiiboSearch
+          this.props.setSearchArray(res);
+        })
+        .catch(err => {
+          console.log("We've got a problem, sir.", err);
+        });
+    }
   };
 
   setID = event => {
-    this.setState({ editID: event.target.value });
+    if (!this.state.deleted) {
+      this.setState({ editID: event.target.value });
+    }
   };
 
   render() {
     return (
       <div>
+        {this.state.deleted
+          ? "Item has been deleted."
+          : "Delete can not be undone!"}
         <form className="form-inline">
           <div className="form-group amiibo-delete">
             <input
               type="text"
-              placeholder="ID (Required to delete)"
+              placeholder="Enter ID to be deleted"
               onChange={this.setID}
               className="form-control"
               style={{ width: "270px" }}
